requirements.txt
anthropic>=0.40.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
rich>=13.0.0
pyTelegramBotAPI>=4.14.0
feedparser>=6.0.0
